chore: improve nvcf_split_benchmark
There are four changes:

1. Obtain Kratos bearer token before running benchmark

If the secret used to obtain the bearer token is invalid, this will not be
discovered until after the benchmark is run, which can take many hours.
Instead, get the bearer token immediately for a fast fail. Because this token
may become invalid before the benchmark is complete, a second call to get the
token is made right before the push.

2. Print the metrics to stdout

If for some reason pushing the cloudevent fails, it will still be visible on
stdin or in logs, which could be useful.

3. Add two additional benchmarking hyperparameters:
  * fp8 weights
  * clip_re_chunk_size

4. Make cloudevent pushing optional

5. Add flag for saving metrics to a local file